[ruby] Move database logic to seperate files for clarity
boot.rb contains Bundler initialization, constant definitions, database
connections and model definitions.

Extracting the database connection and model definition to a separate
db.rb makes it more explicit what belongs where. It also makes it easier
to compare the models and database setup across frameworks.
