+++ frameworks/Ruby/rack-app/config.ru
@@ -1,5 +1,4 @@
 # frozen_string_literal: true
-require_relative 'boot'
 require_relative 'app'
 
 run App
