+++ frameworks/Ruby/rack-sequel/config.ru
@@ -1,4 +1,5 @@
-require_relative 'boot'
+# frozen_string_literal: true
 require_relative 'hello_world'
+
 use Rack::ContentLength
 run HelloWorld.new
