+++ frameworks/Ruby/rack-sequel/db.rb
@@ -1,23 +1,6 @@
 # frozen_string_literal: true
-require "bundler/setup"
-require "time"
-MAX_PK = 10_000
-QUERY_RANGE = (1..MAX_PK).freeze
-ALL_IDS = QUERY_RANGE.to_a
-QUERIES_MIN = 1
-QUERIES_MAX = 500
+
 SEQUEL_NO_ASSOCIATIONS = true
-
-SERVER_STRING = "roda"
-
-Bundler.require(:default) # Load core modules
-
-CONTENT_TYPE = 'Content-Type'
-JSON_TYPE = 'application/json'
-HTML_TYPE = 'text/html; charset=utf-8'
-PLAINTEXT_TYPE = 'text/plain'
-DATE_HEADER = 'Date'
-SERVER_HEADER = 'Server'
 
 def connect(dbtype)
   Bundler.require(dbtype) # Load database-specific modules
@@ -41,21 +24,20 @@ def connect(dbtype)
     opts[:max_connections] = 512
   end
 
-  Sequel.connect "%{adapter}://%{host}/%{database}?user=%{user}&password=%{password}" %
-                   {
-                     adapter: adapter,
-                     host: "tfb-database",
-                     database: "hello_world",
-                     user: "benchmarkdbuser",
-                     password: "benchmarkdbpass"
-                   },
-                 opts
+  Sequel.connect \
+    '%{adapter}://%{host}/%{database}?user=%{user}&password=%{password}' % {
+      adapter: adapter,
+      host: 'tfb-database',
+      database: 'hello_world',
+      user: 'benchmarkdbuser',
+      password: 'benchmarkdbpass'
+    }, opts
 end
 
-DB = connect ENV.fetch("DBTYPE").to_sym
+DB = connect ENV.fetch('DBTYPE').to_sym
 
 # Define ORM models
-class World < Sequel.Model(:World)
+class World < Sequel::Model(:World)
   def_column_alias(:randomnumber, :randomNumber) if DB.database_type == :mysql
 
   def self.batch_update(worlds)
@@ -65,8 +47,8 @@ class World < Sequel.Model(:World)
       ids = []
       sql = String.new("UPDATE world SET randomnumber = CASE id ")
       worlds.each do |world|
-        sql << "when #{world.id} then #{world.randomnumber} "
-        ids << world.id
+        sql << "when #{world[:id]} then #{world[:randomnumber]} "
+        ids << world[:id]
       end
       sql << "ELSE randomnumber END WHERE id IN ( #{ids.join(',')})"
       DB.run(sql)
@@ -74,7 +56,7 @@ class World < Sequel.Model(:World)
   end
 end
 
-class Fortune < Sequel.Model(:Fortune)
+class Fortune < Sequel::Model(:Fortune)
   # Allow setting id to zero (0) per benchmark requirements
   unrestrict_primary_key
 end
