+++ frameworks/Ruby/roda-sequel/hello_world.rb
@@ -1,7 +1,26 @@
 # frozen_string_literal: true
+require 'bundler/setup'
+Bundler.require(:default) # Load core modules
+
+require_relative 'db'
+require 'time'
 
 # Our Rack application to be executed by rackup
 class HelloWorld < Roda
+  MAX_PK = 10_000
+  QUERY_RANGE = (1..MAX_PK).freeze
+  ALL_IDS = QUERY_RANGE.to_a
+  QUERIES_MIN = 1
+  QUERIES_MAX = 500
+
+  CONTENT_TYPE = 'Content-Type'
+  JSON_TYPE = 'application/json'
+  HTML_TYPE = 'text/html; charset=utf-8'
+  PLAINTEXT_TYPE = 'text/plain'
+  DATE_HEADER = 'Date'
+  SERVER_HEADER = 'Server'
+  SERVER_STRING = 'roda'
+
   plugin :hooks
   plugin :render, escape: true, layout_opts: { cache_key: "default_layout" }
   plugin :default_headers, SERVER_HEADER => SERVER_STRING
