+++ frameworks/Ruby/sinatra-sequel/db.rb
@@ -1,27 +1,24 @@
 # frozen_string_literal: true
-require 'bundler/setup'
-require 'time'
 
-MAX_PK = 10_000
-ID_RANGE = (1..MAX_PK).freeze
-ALL_IDS = ID_RANGE.to_a
-QUERIES_MIN = 1
-QUERIES_MAX = 500
 SEQUEL_NO_ASSOCIATIONS = true
-#SERVER_STRING = "Sinatra"
-
-Bundler.require(:default) # Load core modules
 
 def connect(dbtype)
   Bundler.require(dbtype) # Load database-specific modules
 
   opts = {}
 
-  adapter = 'postgresql'
+  if dbtype == :mysql
+    adapter = 'trilogy'
+    opts[:ssl] = true
+    opts[:ssl_mode] = 4 # Trilogy::SSL_PREFERRED_NOVERIFY
+    opts[:tls_min_version] = 3 # Trilogy::TLS_VERSION_12
+  else
+    adapter = 'postgresql'
+  end
 
   # Determine threading/thread pool size and timeout
-  if defined?(Puma) && (threads = Puma.cli_config.options.fetch(:max_threads)) > 1
-    opts[:max_connections] = threads
+  if defined?(Puma)
+    opts[:max_connections] = ENV.fetch('MAX_THREADS')
     opts[:pool_timeout] = 10
   else
     opts[:max_connections] = 512
@@ -37,7 +34,7 @@ def connect(dbtype)
     }, opts
 end
 
-DB = connect 'postgres'
+DB = connect ENV.fetch('DBTYPE').to_sym
 
 # Define ORM models
 class World < Sequel::Model(:World)
