+++ frameworks/Ruby/sinatra/config.ru
@@ -1,3 +1,4 @@
-require_relative 'boot'
+# frozen_string_literal: true
 require_relative 'hello_world'
+
 run HelloWorld.new
