+++ frameworks/Ruby/sinatra/config/puma.rb
@@ -1,0 +1,3 @@
+before_fork do
+  ActiveRecord::Base.connection_handler.clear_active_connections!
+end
